# public/api.yaml
info:
  title: Coronatime API
  description: Here goes description
  version: 0.1.9
servers:
  - url: http://localhost:3000/wane/v1
    description: local server
  - url: http://producation.app/api
    description: production server
openapi: 3.0.0
...
components:
  schemas:
...
    Dette:
      type: object
      properties:
        id:
          type: integer
          description: ID de la dette
        montant:
          type: number
          format: float
          description: Montant de la dette
        client_id:
          type: integer
          description: ID du client
        utilisateur_id:
          type: integer
          description: ID de l'utilisateur
        date:
          type: string
          format: date
          description: Date de la dette
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /dettes:
    get:
      summary: Liste des dettes
      tags:
        - Dette
      parameters:
        - name: solde
          in: query
          required: false
          description: Liste des dettes avec ou sans solde
          schema:
            type: string
            enum:
              - oui
              - non
      responses:
        '200':
          description: Liste des dettes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dette'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Unauthenticated.

    post:
      summary: Créer une nouvelle dette
      description: Cette méthode permet de créer une nouvelle dette en enregistrant les détails de la dette, en mettant à jour les quantités des articles et en enregistrant un paiement associé si nécessaire.
      tags:
        - Dette
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                clientId:
                  type: integer
                  description: ID du client
                montant:
                  type: number
                  format: float
                  description: Montant total de la dette
                paiement:
                  type: object
                  properties:
                    montant:
                      type: number
                      format: float
                      description: Montant du paiement effectué
                articles:
                  type: array
                  items:
                    type: object
                    properties:
                      articleId:
                        type: integer
                        description: ID de l'article
                      qteVente:
                        type: integer
                        description: Quantité vendue de l'article
                      prixVente:
                        type: number
                        format: float
                        description: Prix de vente de l'article
              required:
                - clientId
                - montant
                - paiement
                - articles
      responses:
        '200':
          description: Dette enregistrée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dette'
        '400':
          description: Requête invalide
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Quantité insuffisante pour l'article ID: 1"
        '500':
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Une erreur s'est produite lors de l'enregistrement de la dette.
  /dettes/{id}:
    get:
      summary: Afficher une dette spécifique
      description: Cette méthode permet de récupérer une dette spécifique par son ID.
      tags:
        - Dette
      operationId: showDette
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la dette
          schema:
            type: integer
      responses:
        '200':
          description: Dette trouvée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dette'
        '404':
          description: Dette introuvable
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Dette introuvable
        '500':
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Une erreur s'est produite
...
  /dettes/{id}/articles:
    get:
      summary: Liste des articles d'une dette spécifique
      description: Cette méthode permet de récupérer la liste des articles associés à une dette spécifique par son ID.
      tags:
        - Dette
      operationId: listArticleDette
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la dette
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Liste des articles de la dette récupérée avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      montant:
                        type: number
                        format: float
                        example: 100.00
                      client_id:
                        type: integer
                        example: 2
                      articles:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 10
                            libelle:
                              type: string
                              example: Article Example
                            prix:
                              type: number
                              format: float
                              example: 50.00
                            qteStock:
                              type: integer
                              example: 20
                            qteVente:
                              type: integer
                              example: 5
                            prixVente:
                              type: number
                              format: float
                              example: 50.00
        '403':
          description: Non autorisé
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Vous n'avez pas les autorisations requises
        '404':
          description: Dette introuvable
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Dette introuvable
        '500':
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Une erreur s'est produite
  /dettes/{id}/paiements:
    get:
      summary: Liste des paiements d'une dette spécifique
      description: Cette méthode permet de récupérer la liste des paiements associés à une dette spécifique par son ID.
      tags:
        - Dette
      operationId: listPaiementDette
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la dette
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Liste des paiements de la dette récupérée avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      montant:
                        type: number
                        format: float
                        example: 100.00
                      client_id:
                        type: integer
                        example: 2
                      paiements:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 10
                            montant:
                              type: number
                              format: float
                              example: 50.00
                            date:
                              type: string
                              format: date-time
                              example: '2024-09-03T10:00:00Z'
        '403':
          description: Non autorisé
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Vous n'avez pas les autorisations requises
        '404':
          description: Dette introuvable
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Dette introuvable
        '500':
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Une erreur s'est produite
  /dettes/{id}/paiement:
    post:
      summary: Effectuer un paiement pour une dette spécifique
      description: Cette méthode permet d'enregistrer un paiement pour une dette spécifique par son ID. Vérifie également si le montant restant est supérieur à zéro.
      tags:
        - Dette
      operationId: paiementDette
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la dette
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                montant:
                  type: number
                  format: float
                  example: 50.00
                  description: Montant du paiement à ajouter à la dette
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Paiement effectué avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      montantRestant:
                        type: number
                        format: float
                        example: 50.00
                      montantDu:
                        type: number
                        format: float
                        example: 150.00
                      paiements:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                              example: 10
                            montant:
                              type: number
                              format: float
                              example: 50.00
                            date:
                              type: string
                              format: date-time
                              example: '2024-09-03T10:00:00Z'
        '403':
          description: Non autorisé
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Vous n'avez pas les autorisations requises
        '409':
          description: Conflit
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: La dette est déjà payée ou le montant est supérieur au montant restant
        '404':
          description: Dette introuvable
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Dette introuvable
        '500':
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Une erreur s'est produite